Receiving valid JSON device data Changed some of the flow in kismetDataTool. I also figured out how to get the GPS data I wanted for testing purposes. The best way to do it is by passing a filter with paths. Instead of passing the `kismet.common.location.lat`, pass `kismet.device.base.location/kismet.common.avg_loc/kismet.common.location.log`. I figured this out by looking at a kismet plugins source.
Moving forward I'm going to implement the output methods. A way to pass
the data receivied from the REST api into a kml, csv, etc file format. I
wan't to do this in an intelligent way so that the data output method
can handle the currently unimplemented database portion of this program.